pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "etl_notifier"
version = "0.1.0"
requires-python = ">=3.8"
dependencies = [
    "pyodbc>=4.0.39",
    "PyYAML>=6.0.1",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "azure-identity>=1.23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
]

[project.scripts]
etl-notifier = "etl_notifier.main:main"

[tool.setuptools.packages.find]
where = ["src"]
